--- PersonalScheduler/Application/AppDelegate.swift
@@ -1,38 +1,0 @@
-//
-//  AppDelegate.swift
-//  PersonalScheduler
-//
-//  Created by kjs on 06/01/23.
-//
-
-import UIKit
-import FirebaseCore
-import KakaoSDKCommon
-import FacebookCore
-
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    func application(_ application: UIApplication,
-                     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        FirebaseApp.configure()
-        KakaoSDK.initSDK(appKey: "44bc39fa6aab979297a8e7e1b7a423be")
-        ApplicationDelegate.shared.application(
-            application,
-            didFinishLaunchingWithOptions: launchOptions
-        )
-        return true
-    }
-
-    // MARK: UISceneSession Lifecycle
-    func application(_ application: UIApplication,
-                     configurationForConnecting connectingSceneSession: UISceneSession,
-                     options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        
-        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
-    }
-
-    func application(_ application: UIApplication,
-                     didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {}
-}
-
