+++ PersonalScheduler/Assets.xcassets/AccentColor.colorset/Contents.json
