--- PersonalScheduler/Presentation/LoginScene/Login/View/Subviews/LabelSeparator.swift
@@ -1,59 +1,0 @@
-//
-//  LabelSeparator.swift
-//  PersonalScheduler
-//
-//  Created by 이원빈 on 2023/01/09.
-//
-
-import UIKit
-
-final class LabelSeparator: UIView {
-    
-    private let titleLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.backgroundColor = .systemBackground
-        label.font = .preferredFont(forTextStyle: .body)
-        label.textColor = .systemGray
-        return label
-    }()
-    
-    init(text: String) {
-        super.init(frame: .zero)
-        titleLabel.text = text
-        setupInitialView()
-        addSubviews()
-        setupConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setupInitialView() {
-        backgroundColor = .systemBackground
-    }
-    
-    private func addSubviews() {
-        addSubview(titleLabel)
-    }
-    
-    private func setupConstraints() {
-        NSLayoutConstraint.activate([
-            titleLabel.centerXAnchor.constraint(equalTo: centerXAnchor),
-            titleLabel.centerYAnchor.constraint(equalTo: centerYAnchor),
-            
-            heightAnchor.constraint(equalToConstant: 50)
-        ])
-    }
-    
-    override func draw(_ rect: CGRect) {
-        let separator = UIBezierPath()
-        separator.move(to: CGPoint(x: 0, y: bounds.midY))
-        separator.addLine(to: CGPoint(x: bounds.maxX, y: bounds.midY))
-        separator.lineWidth = 1
-        UIColor.systemGray5.setStroke()
-        separator.stroke()
-        separator.close()
-    }
-}
