--- PersonalScheduler/Presentation/ScheduleScene/ScheduleDetailScene/View/ScheduleDetailViewController.swift
@@ -1,31 +1,0 @@
-//
-//  ScheduleDetailViewController.swift
-//  PersonalScheduler
-//
-//  Created by 이원빈 on 2023/01/09.
-//
-
-import UIKit
-
-final class ScheduleDetailViewController: UIViewController {
-    private let viewModel: ScheduleDetailViewModel
-    
-    init(with viewModel: ScheduleDetailViewModel) {
-        self.viewModel = viewModel
-        super.init(nibName: nil, bundle: nil)
-        setupInitialView()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupInitialView()
-    }
-    
-    private func setupInitialView() {
-        view.backgroundColor = .systemBlue
-    }
-}
