--- PersonalScheduler/Utils/Extension/UITextField+.swift
@@ -1,16 +1,0 @@
-//
-//  UITextField+.swift
-//  PersonalScheduler
-//
-//  Created by 이원빈 on 2023/01/09.
-//
-
-import UIKit
-
-extension UITextField {
-    func addLeftPadding() {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: 10, height: self.frame.height))
-        self.leftView = paddingView
-        self.leftViewMode = .always
-    }
-}
