+++ PersonalScheduler/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  PersonalScheduler
+//
+//  Created by kjs on 06/01/23.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
